requirements.txt
langchain-community
langchain
sqlalchemy-bigquery
google-cloud-bigquery
python-dotenv
pandas
langchain-groq
langchain-openai
langchain-ollama
langgraph
streamlit 
mysql-connector-python
langchainhub
plotly